Remove the core package.
It serves no real purpose. Just a bad pattern from the old MUD.
