+++ src/main/java/com/agonyengine/AgonyEngine.java
@@ -1,0 +1,11 @@
+package com.agonyengine;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AgonyEngine {
+    public static void main(String... args) {
+        SpringApplication.run(AgonyEngine.class, args);
+    }
+}
