+++ src/main/java/com/agonyengine/config/DataSourceConfiguration.java
@@ -1,0 +1,17 @@
+package com.agonyengine.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class DataSourceConfiguration {
+    @Bean
+    @ConfigurationProperties(prefix = "agony.datasource")
+    public DataSource dataSource() {
+        return DataSourceBuilder.create().build();
+    }
+}
