+++ src/main/java/com/agonyengine/config/SessionConfiguration.java
@@ -1,0 +1,9 @@
+package com.agonyengine.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.session.jdbc.config.annotation.web.http.EnableJdbcHttpSession;
+
+@EnableJdbcHttpSession(maxInactiveIntervalInSeconds = 86400)
+@Configuration
+public class SessionConfiguration {
+}
