--- src/main/java/com/agonyengine/core/AgonyEngine.java
@@ -1,11 +1,0 @@
-package com.agonyengine.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AgonyEngine {
-    public static void main(String... args) {
-        SpringApplication.run(AgonyEngine.class, args);
-    }
-}
