--- src/main/java/com/agonyengine/core/config/ApplicationConfiguration.java
@@ -1,21 +1,0 @@
-package com.agonyengine.core.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.Date;
-
-@Configuration
-public class ApplicationConfiguration {
-    private Date bootDate = new Date();
-
-    @Bean(name = "applicationVersion")
-    public String applicationVersion() {
-        return ApplicationConfiguration.class.getPackage().getImplementationVersion();
-    }
-
-    @Bean(name = "applicationBootDate")
-    public Date applicationBootDate() {
-        return bootDate;
-    }
-}
