--- src/main/java/com/agonyengine/core/config/SessionConfiguration.java
@@ -1,9 +1,0 @@
-package com.agonyengine.core.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.session.jdbc.config.annotation.web.http.EnableJdbcHttpSession;
-
-@EnableJdbcHttpSession(maxInactiveIntervalInSeconds = 86400)
-@Configuration
-public class SessionConfiguration {
-}
