--- src/main/java/com/agonyengine/core/config/WebSocketBrokerProperties.java
@@ -1,72 +1,0 @@
-package com.agonyengine.core.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties(prefix = "agony.broker")
-public class WebSocketBrokerProperties {
-    private Boolean ssl;
-    private String host;
-    private Integer port;
-    private String systemUsername;
-    private String systemPassword;
-    private String clientUsername;
-    private String clientPassword;
-
-    public Boolean getSsl() {
-        return ssl;
-    }
-
-    public void setSsl(Boolean ssl) {
-        this.ssl = ssl;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
-    }
-
-    public Integer getPort() {
-        return port;
-    }
-
-    public void setPort(Integer port) {
-        this.port = port;
-    }
-
-    public String getSystemUsername() {
-        return systemUsername;
-    }
-
-    public void setSystemUsername(String systemUsername) {
-        this.systemUsername = systemUsername;
-    }
-
-    public String getSystemPassword() {
-        return systemPassword;
-    }
-
-    public void setSystemPassword(String systemPassword) {
-        this.systemPassword = systemPassword;
-    }
-
-    public String getClientUsername() {
-        return clientUsername;
-    }
-
-    public void setClientUsername(String clientUsername) {
-        this.clientUsername = clientUsername;
-    }
-
-    public String getClientPassword() {
-        return clientPassword;
-    }
-
-    public void setClientPassword(String clientPassword) {
-        this.clientPassword = clientPassword;
-    }
-}
