--- src/main/java/com/agonyengine/core/model/stomp/UserInput.java
@@ -1,13 +1,0 @@
-package com.agonyengine.core.model.stomp;
-
-public class UserInput {
-    private String input;
-
-    public void setInput(String input) {
-        this.input = input;
-    }
-
-    public String getInput() {
-        return input;
-    }
-}
