--- src/main/java/com/agonyengine/core/resource/exception/NoSuchActorException.java
@@ -1,7 +1,0 @@
-package com.agonyengine.core.resource.exception;
-
-public class NoSuchActorException extends Exception {
-    public NoSuchActorException(String message) {
-        super(message);
-    }
-}
