+++ src/main/java/com/agonyengine/model/actor/PlayerActorTemplate.java
@@ -1,0 +1,42 @@
+package com.agonyengine.model.actor;
+
+import org.hibernate.annotations.Type;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.util.UUID;
+
+@Entity
+public class PlayerActorTemplate {
+    @Id
+    @GeneratedValue
+    @Type(type = "pg-uuid")
+    private UUID id;
+    private String account;
+    private String givenName;
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public String getAccount() {
+        return account;
+    }
+
+    public void setAccount(String account) {
+        this.account = account;
+    }
+
+    public String getGivenName() {
+        return givenName;
+    }
+
+    public void setGivenName(String givenName) {
+        this.givenName = givenName;
+    }
+}
