+++ src/main/java/com/agonyengine/model/stomp/UserInput.java
@@ -1,0 +1,13 @@
+package com.agonyengine.model.stomp;
+
+public class UserInput {
+    private String input;
+
+    public void setInput(String input) {
+        this.input = input;
+    }
+
+    public String getInput() {
+        return input;
+    }
+}
