+++ src/main/java/com/agonyengine/resource/exception/NoSuchActorException.java
@@ -1,0 +1,7 @@
+package com.agonyengine.resource.exception;
+
+public class NoSuchActorException extends Exception {
+    public NoSuchActorException(String message) {
+        super(message);
+    }
+}
